requirements.txt
fastapi
uvicorn
pydantic
python-dotenv

streamlit
requests

langchain
langchain-core
langchain-community
langchain-text-splitters

langchain-groq
langchain-huggingface
langchain-tavily

faiss-cpu
sentence-transformers
transformers
huggingface-hub

tavily-python

pypdf